Act_16_ORM/requirements.txt
greenlet==2.0.1
psycopg2-binary==2.9.5
python-decouple==3.6
SQLAlchemy==1.4.44
